requirements.txt
fake_useragent==1.1.3
opencv_python==4.7.0.72
openpyxl==2.5.6
PyAutoGUI==0.9.53
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
schedule==1.2.0
selenium==4.9.1
